# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CIBOL · 实验室系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "lab"
        ],
        "summary": "Update Config",
        "description": "保存当前学期与组会默认（管理员）。无则建、有则更新。",
        "operationId": "update_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/eval/compute": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "Compute",
        "operationId": "compute_api_eval_compute_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalComputeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/eval/rank-series": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "Rank Series",
        "operationId": "rank_series_api_eval_rank_series_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "from_",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2026-04-19",
              "title": "From "
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2026-06-07",
              "title": "To"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "total",
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankSeriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/eval/reports": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "List Reports",
        "description": "评选期内的组会（直接来自 meetings 表，与组会日历同源），供管理员录入。",
        "operationId": "list_reports_api_eval_reports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportOut"
                  },
                  "type": "array",
                  "title": "Response List Reports Api Eval Reports Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/eval/excellence": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "Excellence",
        "operationId": "excellence_api_eval_excellence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcellenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "eval"
        ],
        "summary": "Publish Excellence",
        "operationId": "publish_excellence_api_eval_excellence_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishExcellence"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcellenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/eval/reports/{key}/rating": {
      "post": {
        "tags": [
          "eval"
        ],
        "summary": "Submit Rating",
        "operationId": "submit_rating_api_eval_reports__key__rating_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingSubmit"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/eval/reports/{key}/attendance": {
      "post": {
        "tags": [
          "eval"
        ],
        "summary": "Set Attendance",
        "operationId": "set_attendance_api_eval_reports__key__attendance_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceSet"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/eval/reports/{key}/speaks": {
      "post": {
        "tags": [
          "eval"
        ],
        "summary": "Set Speaks",
        "operationId": "set_speaks_api_eval_reports__key__speaks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeaksSet"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/eval/config": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_eval_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalConfigIO"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "eval"
        ],
        "summary": "Put Config",
        "operationId": "put_config_api_eval_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalConfigIO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalConfigIO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/eval/excellence/all": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "Excellence All",
        "operationId": "excellence_all_api_eval_excellence_all_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExcellenceOut"
                  },
                  "type": "array",
                  "title": "Response Excellence All Api Eval Excellence All Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/eval/reports/{key}/votes": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "List Votes",
        "description": "列出某场组会的全部匿名选票明细，供管理员核查异常评分。",
        "operationId": "list_votes_api_eval_reports__key__votes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoteDetailOut"
                  },
                  "title": "Response List Votes Api Eval Reports  Key  Votes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/eval/votes/{vote_id}": {
      "delete": {
        "tags": [
          "eval"
        ],
        "summary": "Delete Vote",
        "operationId": "delete_vote_api_eval_votes__vote_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vote Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigIn": {
        "properties": {
          "semester": {
            "$ref": "#/components/schemas/SemesterIn"
          },
          "meetingDefault": {
            "$ref": "#/components/schemas/MeetingDefaultIn"
          }
        },
        "type": "object",
        "required": [
          "semester",
          "meetingDefault"
        ],
        "title": "ConfigIn"
      },
      "ConfigOut": {
        "properties": {
          "semester": {
            "$ref": "#/components/schemas/SemesterOut"
          },
          "meetingDefault": {
            "$ref": "#/components/schemas/MeetingDefaultOut"
          }
        },
        "type": "object",
        "required": [
          "semester",
          "meetingDefault"
        ],
        "title": "ConfigOut"
      },
...
      "EvalComputeOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "weights": {
            "additionalProperties": true,
            "type": "object",
            "title": "Weights"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EvalRowOut"
            },
            "type": "array",
            "title": "Rows"
          },
          "merged": {
            "items": {
              "$ref": "#/components/schemas/MergedOut"
            },
            "type": "array",
            "title": "Merged"
          }
        },
        "type": "object",
        "required": [
          "total",
          "weights",
          "rows",
          "merged"
        ],
        "title": "EvalComputeOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankSeriesOut": {
        "properties": {
          "points": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Points"
          },
          "ranks": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Ranks"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "points",
          "ranks",
          "total"
        ],
        "title": "RankSeriesOut"
      },
...
      "SpeaksSet": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "SpeaksSet"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}